Simplify retryWithExpectedResult used in test
Before this change, the function used to evaluated the action multiple times
and to prevent that, one had to evaluate the action once more and pass in the
"first result". This was both unintuitive and uncomfortable, now it only
evaluates action once per retry automatically.

Also removes the feature that there could be more than one assertion. This worked in
a very strange way anyway (recursed for each assertion separately?), was not used much,
and is trivial to overcome by simply putting more assertion into one lambda function.

This commit should fix flakyness of some tests (lookup by dataSetId in particular)