Make limit consistent
The scala sdk regards the limit to be the total amount of rows returned in some cases, while in other cases it regards it as the batch size.
This commit makes limit consistently reflect the total amount of items to return for filter. There is still a logic bug in read without filter which will be fixed by another commit.