Update to Cats Effect 3, drop Scala 2.11
Also drops retrying with the Id monad. We could add it back with a
super hacky implementation of Temporal that only implements `sleep`,
but Id is really only meant for testing and demos, where retries
shouldn't be important.

The actual tests in this repository can use IO and unsafeRunSync when
testing with retries.