Expose API for partitioned filtering
In Spark Data Source we currently use the filterPartitionsF function,
but it is not ideal as it requests new set of cursors on each
executor. I'd like to change the logic so we request cursors on the driver and then
only use the filter function on executors to make sure
that we use the same set of cursors everywhere.