Use TestControl for better testing of CachedResource, as for CE2
When upgrading to cats-effect 3 I was lazy about this and switched it
to use real IO and sleep, which made the tests slow and flaky.

TestControl in cats-effect 3 replaces TestContext and mocked Timer in
cats-effect 2, and has a much nicer interface.

So now the tests run instantly, and should no longer be flaky.