Add support for the well data layer (#587)
This commit adds support for the well-data-layer for the purpose of
adding it to transformations.

It has a specific part for ingesting sources, wells, wellbores, and
setting merge rules, and a generic part that allows you to ingest an
`Items[JsonObject]` and retrieve `ItemsWithCursor[JsonObject]`. The
specific parts are going to be used only for tests, while the generic
part will be used for the actual transformations.

This approach was chosen so that we wouldn't have to update any
SDK's if we decide to extend the well data layer data model.