Drop .unsafeRunSync from main code (RateLimitingBackend)
It can be shifted to the caller for more safety.
Excessive unsafeRunSync, especially if some of actions aren't wrapped
as `blocking` may lead to deadlocks.

For RateLimitingBackend case the more generic overload should work
and caller is free to have unsafeRunSync on top if not ready to use
it as effect.

Should be the last remaining `.unsafeRun*` in `src/main` in this repo