Expose filtering fields with null values from Raw (#798)
* Expose filtering fields with null values from Raw

This exposes the query parameter to control whether Raw should return
fields where the value is null or not. After several failed tries, it
does this through the RawRows constructor, as this is also relevant
for clients that do not filter the output (so it can't just be part of
RawRowsFilter, sadly). The filter is simply not a query filter, like
the other components of RawRowsFilter.

* wartignore

---------

Co-authored-by: Dmitry Ivankov <dmitry.ivankov@cognite.com>